shellrc: libvirt vm now boot with UEFI instead of BIOS

+++ .oh-my-shell/shellrc/plugins/libvirt/functions.sh
@@ -36,7 +36,7 @@ function f_libvirt_pool_local_setup {
             --pool "$poolName"
 }
 
-function f_libvirt_pool_volume_define {
+function f_libvirt_pool_volume_disk_define {
     local poolName="$1"
     local volName="$2"
     local volFormat="$3"
@@ -67,6 +67,30 @@ function f_libvirt_pool_volume_define {
 
     mv "$srcFile" "$dstFile"
 }
+
+function f_libvirt_pool_volume_cdrom_define {
+    local poolName="$1"
+    local volName="$2"
+
+    #shellcheck disable=SC2086,SC2155
+    local poolPath="$(virsh pool-dumpxml --pool $poolName --xpath '//pool/target/path/text()')"
+
+    local srcFile="$poolPath/$volName.final"
+    local dstFile="$poolPath/$volName"
+
+    #shellcheck disable=SC2086
+    virsh ${VIRSH_OPTIONS} vol-delete \
+        --pool "$poolName" \
+        --vol "$volName"
+
+    #shellcheck disable=SC2086
+    virsh ${VIRSH_OPTIONS} vol-create-as \
+        --pool "$poolName" \
+        --name "$volName" \
+        --capacity "700M" \
+        --format "raw"
+}
+ 
  
 
 function f_libvirt_domain_start {
@@ -81,9 +105,11 @@ function f_libvirt_domain_start {
     vncURL="vnc://localhost$(virsh ${VIRSH_OPTIONS} vncdisplay --domain $domainName)"
     # FIXME: Deal with non ubuntu os later
     sshURL="ssh ubuntu@localhost -p 222x"
+    consoleURL="virsh console $domainName + ENTER + type user login's password"
 
     echo "Access methods:"
     echo "VNC: $vncURL"
+    echo "Console: $consoleURL"
     echo "SSH(maybe): $sshURL"
 }
 
@@ -115,6 +141,7 @@ function f_libvirt_domain_define {
     #shellcheck disable=SC2086
     virsh ${VIRSH_OPTIONS} define \
             --file "$domainXMLFile" \
+        && virsh ${VIRSH_OPTIONS} dumpxml "$domainName" \
         && virsh ${VIRSH_OPTIONS} domuuid "$domainName"
 }
 
@@ -126,8 +153,9 @@ function f_libvirt_domain_setup {
     local domainName="$(xmllint --xpath '//domain/name/text()' $domainXMLFile)"
 
     f_libvirt_pool_local_setup "$poolName" "$poolPath" \
-        && f_libvirt_pool_volume_define "$poolName" "disk-os" "qcow2" "20G" \
-        && f_libvirt_pool_volume_define "$poolName" "disk-data0" "qcow2" "50G" \
+        && f_libvirt_pool_volume_disk_define "$poolName" "disk-os" "qcow2" "20G" \
+        && f_libvirt_pool_volume_disk_define "$poolName" "disk-data0" "qcow2" "50G" \
+        && f_libvirt_pool_volume_cdrom_define "$poolName" "disk-cdrom" \
         && f_libvirt_domain_define "$domainName" "$domainXMLFile"
 }
 
@@ -135,13 +163,18 @@ function f_libvirt_domain_setup_simple {
     local domainName="$1"
     local domainXMLDir="$2"
     local domainTemplateXMLFile="$3"
-    local hostForward="$4"
+    local vcpu="$4"
+    local memory="$5"
+    local hostForward="$6"
 
     # Duplicate template and edit it
     local domainXMLFile="$domainXMLDir/${domainName}.xml"
     mkdir -p "$domainXMLDir"
     xmlstarlet ed -u "//domain/name" -v "$domainName" "$domainTemplateXMLFile" > "$domainXMLFile"
+    sed -i "s/__VCPU__/$vcpu/g" "$domainXMLFile"
+    sed -i "s/__MEMORY__/$memory/g" "$domainXMLFile"
     sed -i "s/__HOSTFWD__/$hostForward/g" "$domainXMLFile"
+    sed -i "s/__POOLNAME__/$domainName/g" "$domainXMLFile"
 
     #shellcheck disable=SC2155,SC2086
     local domainName="$(xmllint --xpath '//domain/name/text()' $domainXMLFile)"
@@ -173,13 +206,19 @@ function f_libvirt_domain_unsetup {
         "$domainName"
 }
 
-function f_libvirt_domain_setup_ubuntu_2204 {
-    local domainName="$1"
-    local hostForward="$2"
 
+function f_libvirt_domain_setup_ubuntu_generic {
+    local releaseName="$1"
+    local domainName="$2"
+    local desktopMode="${3:-1}"
     #Note: Don't use image with suffix '-kvm.img'. These don't boot
-    local imgURL="https://cloud-images.ubuntu.com/jammy/current/jammy-server-cloudimg-amd64.img"
-    local imgOSOriginalFile="/tmp/$$.jammy.server.ubuntu.amd64.img"
+    #eg: "https://cloud-images.ubuntu.com/focal/current/focal-server-cloudimg-amd64.img"
+    local imgURL="$4"
+    local imgLiveCDURL="$5"
+    local hostForward="$6"
+
+    local imgOSOriginalFile="/tmp/$$.${releaseName}.server.ubuntu.amd64.img"
+    local imgOSLiveOriginalFile="/tmp/$$.${releaseName}.desktop.ubuntu.amd64.iso"
     local imgSeedFile="${imgOSOriginalFile}.seed.img"
     local imgFile="${imgOSOriginalFile}.qcow2"
     local userDataFile="${imgFile}.user_data"
@@ -191,7 +230,6 @@ function f_libvirt_domain_setup_ubuntu_2204 {
     # https://askubuntu.com/a/1081171/226227
     # cloud-init options: https://cloudinit.readthedocs.io/en/latest/topics/modules.html
 
-#username: ubuntu
 cat > "$userDataFile" <<EOF
 #cloud-config
 password: ubuntu
@@ -201,7 +239,16 @@ EOF
 
     echo "instance-id: $(uuidgen || echo i-abcdefg)" > "$metaDataFile"
 
-    wget --no-timestamping -O "$imgOSOriginalFile" -nc "$imgURL"
+    if [[ "$desktopMode" == "0" ]];
+    then
+        wget --no-timestamping -O "$imgOSLiveOriginalFile" -nc "$imgLiveCDURL"
+        vcpu=4
+        memory=4
+    else
+        wget --no-timestamping -O "$imgOSOriginalFile" -nc "$imgURL"
+        vcpu=2
+        memory=2
+    fi
 
     # Build + run cloud-locals to generate seed.img containing cloud-config settings
     docker build . \
@@ -214,21 +261,49 @@ EOF
     docker run \
         -t \
         --rm \
-        --mount type=bind,source=${imgSeedFile},target=/mnt/seed.img \
-        --mount type=bind,source=${userDataFile},target=/mnt/user_data \
-        --mount type=bind,source=${metaDataFile},target=/mnt/meta_data \
+        --mount type=bind,source="${imgSeedFile}",target="/mnt/seed.img" \
+        --mount type=bind,source="${userDataFile}",target="/mnt/user_data" \
+        --mount type=bind,source="${metaDataFile}",target="/mnt/meta_data" \
         "$containerTag" \
         cloud-localds -v /mnt/seed.img /mnt/user_data /mnt/meta_data
 
     # Create VM
     local domainXMLTemplateFile="$LIBVIRT_THIS_PLUGIN_DIR/image-builder/ubuntu/domain-template.xml"
     local domainXMLDir="$PWD/${domainName}"
-    f_libvirt_domain_setup_simple "$domainName" "$domainXMLDir" "$domainXMLTemplateFile" "$hostForward"
+    f_libvirt_domain_setup_simple "$domainName" "$domainXMLDir" "$domainXMLTemplateFile" "$vcpu" "$memory" "$hostForward"
 
-    # Copy os + seed
-    cp "$imgOSOriginalFile" "$domainXMLDir/disk-os"
-    cp "$imgSeedFile" "$domainXMLDir/disk-seed"
+    # Copy os + seed disks
+    [ -f "$imgOSOriginalFile" ] && cp "$imgOSOriginalFile" "$domainXMLDir/disk-os" && qemu-img resize "$domainXMLDir/disk-os" +20G
+    [ -f "$imgSeedFile" ] && cp "$imgSeedFile" "$domainXMLDir/disk-seed"
+    [ -f "$imgOSLiveOriginalFile" ] && [[ "$desktopMode" == "0" ]] && cp "$imgOSLiveOriginalFile" "$domainXMLDir/disk-cdrom"
+}
 
-    # Resize os disk size
-    qemu-img resize "$domainXMLDir/disk-os" +20G
+
+function f_libvirt_domain_setup_ubuntu_2204 {
+    local domainName="$1"
+    local desktopMode="$2"
+    local hostForward="$3"
+
+    f_libvirt_domain_setup_ubuntu_generic \
+        "jammy" \
+        "$domainName" \
+        "$desktopMode" \
+        "https://cloud-images.ubuntu.com/jammy/current/jammy-server-cloudimg-amd64.img" \
+        "https://releases.ubuntu.com/22.04/ubuntu-22.04.1-desktop-amd64.iso" \
+        "$hostForward"
+}
+
+
+function f_libvirt_domain_setup_ubuntu_2004 {
+    local domainName="$1"
+    local desktopMode="$2"
+    local hostForward="$3"
+
+    f_libvirt_domain_setup_ubuntu_generic \
+        "focal" \
+        "$domainName" \
+        "$desktopMode" \
+        "https://cloud-images.ubuntu.com/focal/current/focal-server-cloudimg-amd64.img" \
+        "https://releases.ubuntu.com/20.04/ubuntu-20.04.5-desktop-amd64.iso" \
+        "$hostForward"
 }
